src/ph2-qlora-finetunning/requirements.txt
torch
transformers
datasets
peft
bitsandbytes
huggingface_hub
python-dotenv
